% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MAX_MIN_rcpp}
\alias{MAX_MIN_rcpp}
\title{Comparacion entre maximos y minimos en Matrices cuadradas}
\usage{
MAX_MIN_rcpp(A, B)
}
\arguments{
\item{NumericMatrix}{B : Matriz cuadrada de NxN}
}
\value{
Devuelve una matriz de NxN con los valores resultantes de dicha comparacion
realizada dentro de la funcion.
}
\description{
Compara la primera fila de la primera matriz con la columna de la
segunda matriz tomando los valores minimos entre ambos y generando un nuevo vector
luego dentro de ese vector ver cual es el maximo, lo que nos lleva al valor que queda
en esa posicion
}
\examples{
matrizPrueba1 <- matrix(round(runif(16, 1, 10)), nrow = 4, ncol = 4);
matrizPrueba2 <- matrix(round(runif(16, 10, 20)), nrow = 4, ncol = 4);
set.seed(123);
matrizPrueba1;
matrizPrueba2;
MAX_MIN_rcpp(matrizPrueba1, matrizPrueba2);
}
