% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiMATRIZ_rcpp}
\alias{multiMATRIZ_rcpp}
\title{Multiplicacion de Objetos R de tipo Matriz}
\usage{
multiMATRIZ_rcpp(A, B)
}
\arguments{
\item{NumericMatrix}{B : MAtriz de dimension MxN}
}
\value{
Devuelve una matriz de resultado de dimension N filas de primera
matriz y N columnas de segunda matriz.
}
\description{
La funcion realiza la multiplicacion de dos matrices
mediante su formulacion matematica de multiplicacion de matrices, lo que
quiere decir que es necesario que el numero de filas de la primera matriz
sea igual al numero de columnas de la segunda matriz
}
\examples{
matrizPrueba1 <- matrix(round(runif(16, 1, 10)), nrow = 4, ncol = 4);
matrizPrueba2 <- matrix(round(runif(16, 10, 20)), nrow = 4, ncol = 4);
set.seed(123);
matrizPrueba1;
matrizPrueba2;
multiMATRIZ_rcpp(matrizPrueba1, matrizPrueba2);
}
