% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rBIND_rcpp}
\alias{rBIND_rcpp}
\title{Combinar Objetos R por filas}
\usage{
rBIND_rcpp(A, vect)
}
\arguments{
\item{RObjeto}{A : Objeto que puede ser un DataFrame o Matriz.}

\item{NumericVector}{x : Objeto de tipo vector numerico que debe ser del mismo tamaño
que las columnas del DataFrame o Matriz.}
}
\value{
Devuelve una Matriz con el vector añadido como nueva Fila en caso que se le pase una matriz como
parametro y un DataFrame con el vector como nueva Fila en caso que se le pase un
DataFrame como parametro.
}
\description{
Funcion que realiza la union de un DataFrame o Matriz con un vector.
}
\examples{
x <- c(2,3,4,5)
x
matrizPrueba1 <- matrix(round(runif(16, 1, 10)), nrow = 4, ncol = 4);
set.seed(123);
matrizPrueba1;
mi_df <- data.frame(1:4, 5:8, 9:12, 13:16);
mi_df;
rBIND_rcpp(mi_df, x);
rBIND_rcpp(matrizPrueba1, x);
}
