% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whichMIN_rcpp}
\alias{whichMIN_rcpp}
\title{Buscardor de indices de maximos en Objetos de R}
\usage{
whichMIN_rcpp(x)
}
\arguments{
\item{RObjeto}{A : Objeto R para la realizacion de la busqueda de indice de minimo.}
}
\value{
Devuelve un DataFrame conteniendo en sus columnas las posiciones en las que
se encuentran el o los valores minimos dentro del objeto.
}
\description{
Funcion que realiza la busqueda de el o los minimos dentro de
un objeto de R.
}
\examples{
matrizPrueba1 <- matrix(round(runif(16, 1, 10)), nrow = 4, ncol = 4);
set.seed(123);
matrizPrueba1;
whichMIN_rcpp(matrizPrueba1);
}
